Anuj/Assignment 1/requirements.txt
beautifulsoup4==4.13.3
psycopg2_binary==2.9.10
Requests==2.32.3
